+++ lib/consts/consts.go
@@ -9,6 +9,11 @@ const MaxInsertPacketSizeForVMStorage = 100 * 1024 * 1024
 // max memory usage occupied by buffers at vminsert and vmstorage.
 const MaxInsertPacketSizeForVMInsert = 30 * 1024 * 1024
 
+// MinInsertPacketSizeForVMInsert is the border when setting maxBufSizePerStorageNode, which is calculated based on various factors.
+// It means maxBufSizePerStorageNode should not be lower than MinInsertPacketSizeForVMInsert, but the actual request size could be smaller,
+// if there's only limited amount of data.
+const MinInsertPacketSizeForVMInsert = 1 * 1024 * 1024
+
 // StorageStatusAck defines status response from vmstorage which indicates that request
 // was successfully processed
 // used in legacy streaming protocol (prior RPC)